Move `FishFight::PublicInfo` into it's own file
